Split plugin tab and background tab
