+++ frontend/src/features/background-manager/BackgroundManagerPage.tsx
@@ -1,19 +1,6 @@
-import { memo, useCallback, useEffect, useMemo, useState } from "react";
-import type { ChangeEvent, UIEvent } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import {
-  ArrowDown,
-  ArrowUp,
-  Download,
-  ExternalLink,
-  Image as ImageIcon,
-  Languages,
-  Music,
-  Plus,
-  Save,
-  Trash2,
-  Upload,
-} from "lucide-react";
+import { Download, ExternalLink, Languages, Plus, Save, Trash2, Upload } from "lucide-react";
 
 import {
   backgroundsQueryKey,
@@ -33,7 +20,7 @@ import {
   uploadBackgroundImages,
 } from "../../entities/background/repository";
 import type { Background } from "../../entities/config/types";
-import { fileUrl, openExternal } from "../../entities/files/repository";
+import { openExternal } from "../../entities/files/repository";
 import { baseName, numberedTags, tagContents } from "../../shared/assets/assetText";
 import { useI18n } from "../../shared/i18n";
 import {
@@ -42,275 +29,21 @@ import {
   Button,
   EmptyState,
   FilePicker,
-  ImageAssetGallery,
-  PathDisplay,
   QueryErrorState,
   TextArea,
   TextInput,
   useToast,
 } from "../../shared/ui";
+import { BackgroundMusicSection } from "./BackgroundMusicSection";
+import { BackgroundSpriteGallery } from "./BackgroundSpriteGallery";
+import {
+  createBackground,
+  type BackgroundBgmItem,
+  type BackgroundDeleteTarget,
+  type BgmSortDirection,
+  type BgmSortKey,
+} from "./backgroundUtils";
 import "./BackgroundManagerPage.css";
-
-function createBackground(): Background {
-  return {
-    bg_tags: "",
-    bgm_list: [],
-    bgm_tags: "",
-    name: "",
-    sprite_prefix: "temp",
-    sprites: [],
-  };
-}
-
-const BGM_ROW_HEIGHT = 58;
-const VIRTUAL_OVERSCAN_ROWS = 4;
-const VIRTUAL_BGM_ROWS = 10;
-
-type BackgroundDeleteTarget =
-  | { kind: "background"; name: string }
-  | { filename: string; index: number; kind: "image"; name: string }
-  | { count: number; kind: "all-images"; name: string }
-  | { filename: string; index: number; kind: "bgm"; name: string }
-  | { count: number; indexes: number[]; kind: "selected-bgm"; name: string }
-  | { count: number; kind: "all-bgm"; name: string };
-
-type BgmSortDirection = "asc" | "desc";
-type BgmSortKey = "filename" | "index";
-
-interface BackgroundBgmItem {
-  filename: string;
-  originalIndex: number;
-  path: string;
-}
-
-function useVirtualRange(count: number, rowHeight: number, visibleRows: number) {
-  const [scrollTop, setScrollTop] = useState(0);
-  const viewportHeight = Math.max(rowHeight, Math.min(count || 1, visibleRows) * rowHeight);
-  const maxScrollTop = Math.max(0, count * rowHeight - viewportHeight);
-  const clampedScrollTop = Math.min(scrollTop, maxScrollTop);
-  const startIndex = Math.max(0, Math.floor(clampedScrollTop / rowHeight) - VIRTUAL_OVERSCAN_ROWS);
-  const endIndex = Math.min(count, startIndex + visibleRows + VIRTUAL_OVERSCAN_ROWS * 2);
-  const onScroll = useCallback((event: UIEvent<HTMLElement>) => {
-    setScrollTop(event.currentTarget.scrollTop);
-  }, []);
-
-  return {
-    endIndex,
-    maxHeight: count > visibleRows ? viewportHeight : undefined,
-    onScroll,
-    paddingBottom: Math.max(0, (count - endIndex) * rowHeight),
-    paddingTop: startIndex * rowHeight,
-    startIndex,
-  };
-}
-
-interface BackgroundBgmRowsProps {
-  allSelected: boolean;
-  clearSelectionLabel: string;
-  deleting: boolean;
-  filenameLabel: string;
-  items: BackgroundBgmItem[];
-  indexLabel: string;
-  onDelete: (index: number) => void;
-  onSort: (key: BgmSortKey) => void;
-  onToggleAllSelection: () => void;
-  onTagChange: (index: number, value: string) => void;
-  onToggleSelection: (index: number, checked: boolean) => void;
-  pathLabel: string;
-  previewLabel: string;
-  removeLabel: string;
-  rowTags: string[];
-  selectLabel: string;
-  selectAllLabel: string;
-  selectedIndexes: Set<number>;
-  sortDirection: BgmSortDirection;
-  sortKey: BgmSortKey;
-  tagLabel: string;
-}
-
-interface BackgroundBgmRowProps {
-  deleting: boolean;
-  index: number;
-  onDelete: (index: number) => void;
-  onTagChange: (index: number, value: string) => void;
-  onToggleSelection: (index: number, checked: boolean) => void;
-  path: string;
-  removeLabel: string;
-  selected: boolean;
-  tag: string;
-}
-
-const BackgroundBgmRow = memo(function BackgroundBgmRow({
-  deleting,
-  index,
-  onDelete,
-  onTagChange,
-  onToggleSelection,
-  path,
-  removeLabel,
-  selected,
-  tag,
-}: BackgroundBgmRowProps) {
-  const handleDelete = useCallback(() => onDelete(index), [index, onDelete]);
-  const handleTagChange = useCallback(
-    (event: ChangeEvent<HTMLInputElement>) => onTagChange(index, event.target.value),
-    [index, onTagChange],
-  );
-  const handleToggle = useCallback(
-    (event: ChangeEvent<HTMLInputElement>) => onToggleSelection(index, event.target.checked),
-    [index, onToggleSelection],
-  );
-  const filename = baseName(path);
-
-  return (
-    <tr aria-selected={selected}>
-      <td>
-        <input checked={selected} onChange={handleToggle} type="checkbox" />
-      </td>
-      <td>{index + 1}</td>
-      <td className="background-bgm-table__filename" title={filename}>
-        <span className="background-bgm-table__filename-inner">
-          <Music aria-hidden className="asset-row__icon" />
-          <span>{filename}</span>
-        </span>
-      </td>
-      <td>
-        <PathDisplay className="background-bgm-table__path" path={path} />
-      </td>
-      <td className="background-bgm-table__tag">
-        <TextInput onChange={handleTagChange} value={tag} />
-      </td>
-      <td className="background-bgm-table__preview">
-        {path ? <audio className="audio-inline" controls preload="none" src={fileUrl(path)} /> : null}
-      </td>
-      <td>
-        <AsyncButton
-          icon={<Trash2 aria-hidden className="button__icon" />}
-          loading={deleting}
-          onClick={handleDelete}
-          variant="ghost"
-        >
-          {removeLabel}
-        </AsyncButton>
-      </td>
-    </tr>
-  );
-});
-
-function BgmSpacerRow({ height }: { height: number }) {
-  if (!height) {
-    return null;
-  }
-  return (
-    <tr aria-hidden className="background-virtual-table__spacer">
-      <td colSpan={7} style={{ height }} />
-    </tr>
-  );
-}
-
-const BackgroundBgmRows = memo(function BackgroundBgmRows({
-  allSelected,
-  clearSelectionLabel,
-  deleting,
-  filenameLabel,
-  items,
-  indexLabel,
-  onDelete,
-  onSort,
-  onToggleAllSelection,
-  onTagChange,
-  onToggleSelection,
-  pathLabel,
-  previewLabel,
-  removeLabel,
-  rowTags,
-  selectLabel,
-  selectAllLabel,
-  selectedIndexes,
-  sortDirection,
-  sortKey,
-  tagLabel,
-}: BackgroundBgmRowsProps) {
-  const virtual = useVirtualRange(items.length, BGM_ROW_HEIGHT, VIRTUAL_BGM_ROWS);
-  const visibleItems = items.slice(virtual.startIndex, virtual.endIndex);
-  const indexAriaSort = sortKey === "index" ? (sortDirection === "asc" ? "ascending" : "descending") : undefined;
-  const filenameAriaSort = sortKey === "filename" ? (sortDirection === "asc" ? "ascending" : "descending") : undefined;
-  const SortIcon = sortDirection === "asc" ? ArrowUp : ArrowDown;
-
-  return (
-    <div
-      className="data-table-wrap background-virtual-table"
-      onScroll={virtual.onScroll}
-      style={{ maxHeight: virtual.maxHeight }}
-    >
-      <table className="data-table background-bgm-table">
-        <colgroup>
-          <col className="background-bgm-table__select-col" />
-          <col className="background-bgm-table__index-col" />
-          <col className="background-bgm-table__filename-col" />
-          <col className="background-bgm-table__path-col" />
-          <col className="background-bgm-table__tag-col" />
-          <col className="background-bgm-table__preview-col" />
-          <col className="background-bgm-table__remove-col" />
-        </colgroup>
-        <thead>
-          <tr>
-            <th>
-              <button
-                aria-label={allSelected ? clearSelectionLabel : selectAllLabel}
-                aria-pressed={allSelected}
-                className="background-bgm-table__header-button"
-                onClick={onToggleAllSelection}
-                title={allSelected ? clearSelectionLabel : selectAllLabel}
-                type="button"
-              >
-                {selectLabel}
-              </button>
-            </th>
-            <th aria-sort={indexAriaSort}>
-              <button className="background-bgm-table__header-button" onClick={() => onSort("index")} type="button">
-                <span>{indexLabel}</span>
-                {sortKey === "index" ? <SortIcon aria-hidden className="background-bgm-table__sort-indicator" /> : null}
-              </button>
-            </th>
-            <th aria-sort={filenameAriaSort}>
-              <button className="background-bgm-table__header-button" onClick={() => onSort("filename")} type="button">
-                <span>{filenameLabel}</span>
-                {sortKey === "filename" ? (
-                  <SortIcon aria-hidden className="background-bgm-table__sort-indicator" />
-                ) : null}
-              </button>
-            </th>
-            <th>{pathLabel}</th>
-            <th>{tagLabel}</th>
-            <th>{previewLabel}</th>
-            <th>{removeLabel}</th>
-          </tr>
-        </thead>
-        <tbody>
-          <BgmSpacerRow height={virtual.paddingTop} />
-          {visibleItems.map((item) => {
-            return (
-              <BackgroundBgmRow
-                deleting={deleting}
-                index={item.originalIndex}
-                key={`${item.path}-${item.originalIndex}`}
-                onDelete={onDelete}
-                onTagChange={onTagChange}
-                onToggleSelection={onToggleSelection}
-                path={item.path}
-                removeLabel={removeLabel}
-                selected={selectedIndexes.has(item.originalIndex)}
-                tag={rowTags[item.originalIndex] ?? ""}
-              />
-            );
-          })}
-          <BgmSpacerRow height={virtual.paddingBottom} />
-        </tbody>
-      </table>
-    </div>
-  );
-});
 
 export function BackgroundManagerPage() {
   const queryClient = useQueryClient();
@@ -335,10 +68,11 @@ export function BackgroundManagerPage() {
   const [nameError, setNameError] = useState("");
 
   const selected = useMemo(
-    () => (isCreating ? undefined : (data.find((background) => background.name === selectedName) ?? data[0])),
+    () => (isCreating ? undefined : (data.find((bg) => bg.name === selectedName) ?? data[0])),
     [data, isCreating, selectedName],
   );
   const currentBackgroundName = isCreating ? "" : selectedName;
+
   useEffect(() => {
     if (selected) {
       setSelectedName(selected.name);
@@ -462,11 +196,7 @@ export function BackgroundManagerPage() {
 
   const imageUploadMutation = useMutation({
     mutationFn: () =>
-      uploadBackgroundImages({
-        bgTags: draft.bg_tags,
-        name: currentBackgroundName,
-        paths: pendingImagePaths,
-      }),
+      uploadBackgroundImages({ bgTags: draft.bg_tags, name: currentBackgroundName, paths: pendingImagePaths }),
     onError(error) {
       showToast({
         kind: "error",
@@ -500,11 +230,7 @@ export function BackgroundManagerPage() {
 
   const bgmUploadMutation = useMutation({
     mutationFn: () =>
-      uploadBackgroundBgm({
-        bgmTags: draft.bgm_tags,
-        name: currentBackgroundName,
-        paths: pendingBgmPaths,
-      }),
+      uploadBackgroundBgm({ bgmTags: draft.bgm_tags, name: currentBackgroundName, paths: pendingBgmPaths }),
     onError(error) {
       showToast({
         kind: "error",
@@ -693,6 +419,7 @@ export function BackgroundManagerPage() {
     },
     [currentBackgroundName, draft.sprites, showToast, t],
   );
+
   const handleBgmDelete = useCallback(
     (index: number) => {
       const path = draft.bgm_list[index];
@@ -751,10 +478,7 @@ export function BackgroundManagerPage() {
                   name: pendingDelete.name,
                 })
               : pendingDelete.kind === "all-images"
-                ? t("background.asset.clearImagesConfirmBody", {
-                    count: pendingDelete.count,
-                    name: pendingDelete.name,
-                  })
+                ? t("background.asset.clearImagesConfirmBody", { count: pendingDelete.count, name: pendingDelete.name })
                 : pendingDelete.kind === "bgm"
                   ? t("background.asset.deleteBgmConfirmBody", {
                       filename: pendingDelete.filename,
@@ -766,10 +490,7 @@ export function BackgroundManagerPage() {
                         count: pendingDelete.count,
                         name: pendingDelete.name,
                       })
-                    : t("background.asset.clearBgmConfirmBody", {
-                        count: pendingDelete.count,
-                        name: pendingDelete.name,
-                      }),
+                    : t("background.asset.clearBgmConfirmBody", { count: pendingDelete.count, name: pendingDelete.name }),
         confirmLabel:
           pendingDelete.kind === "image" || pendingDelete.kind === "bgm" ? t("common.remove") : t("common.delete"),
         title:
@@ -801,47 +522,21 @@ export function BackgroundManagerPage() {
     });
   };
 
-  const bgmRowTags = useMemo(
-    () => tagContents(draft.bgm_tags, draft.bgm_list.length),
-    [draft.bgm_list.length, draft.bgm_tags],
-  );
+  const bgmRowTags = useMemo(() => tagContents(draft.bgm_tags, draft.bgm_list.length), [draft.bgm_list.length, draft.bgm_tags]);
   const sortedBgmItems = useMemo(() => {
     const direction = bgmSort.direction === "asc" ? 1 : -1;
     return draft.bgm_list
-      .map((path, originalIndex) => ({
-        filename: baseName(path),
-        originalIndex,
-        path,
-      }))
+      .map((path, originalIndex) => ({ filename: baseName(path), originalIndex, path }))
       .sort((left, right) => {
         if (bgmSort.key === "filename") {
-          const filenameOrder = left.filename.localeCompare(right.filename, undefined, {
-            numeric: true,
-            sensitivity: "base",
-          });
+          const filenameOrder = left.filename.localeCompare(right.filename, undefined, { numeric: true, sensitivity: "base" });
           return (filenameOrder || left.originalIndex - right.originalIndex) * direction;
         }
         return (left.originalIndex - right.originalIndex) * direction;
       });
   }, [bgmSort.direction, bgmSort.key, draft.bgm_list]);
-  const imageRowTags = useMemo(
-    () => tagContents(draft.bg_tags, draft.sprites.length),
-    [draft.bg_tags, draft.sprites.length],
-  );
-  const selectedImage = draft.sprites[selectedImageIndex];
-  const backgroundImageItems = useMemo(
-    () =>
-      draft.sprites.map((sprite, index) => ({
-        id: `${sprite.path}-${index}`,
-        imageSrc: sprite.path ? fileUrl(sprite.path) : "",
-        meta: imageRowTags[index] || "",
-        title: baseName(sprite.path) || `${index + 1}`,
-      })),
-    [draft.sprites, imageRowTags],
-  );
+  const imageRowTags = useMemo(() => tagContents(draft.bg_tags, draft.sprites.length), [draft.bg_tags, draft.sprites.length]);
   const selectedBgmIndexSet = useMemo(() => new Set(selectedBgmIndexes), [selectedBgmIndexes]);
-  const selectedBgmCount = selectedBgmIndexes.filter((index) => index >= 0 && index < draft.bgm_list.length).length;
-  const allBgmSelected = draft.bgm_list.length > 0 && selectedBgmCount === draft.bgm_list.length;
 
   return (
     <div className="page">
@@ -885,9 +580,7 @@ export function BackgroundManagerPage() {
             disabled={!pendingImportFiles.length}
             icon={<Upload aria-hidden className="button__icon" />}
             loading={importMutation.isPending}
-            onClick={() => {
-              importMutation.mutate(pendingImportFiles);
-            }}
+            onClick={() => importMutation.mutate(pendingImportFiles)}
           >
             {t("common.import")}
           </AsyncButton>
@@ -896,11 +589,7 @@ export function BackgroundManagerPage() {
             loading={exportMutation.isPending}
             onClick={() => {
               if (!currentBackgroundName) {
-                showToast({
-                  kind: "error",
-                  message: t("background.validation.nameRequired"),
-                  title: t("common.export"),
-                });
+                showToast({ kind: "error", message: t("background.validation.nameRequired"), title: t("common.export") });
                 return;
               }
               exportMutation.mutate(currentBackgroundName);
@@ -908,12 +597,7 @@ export function BackgroundManagerPage() {
           >
             {t("common.export")}
           </AsyncButton>
-          <AsyncButton
-            icon={<Save aria-hidden className="button__icon" />}
-            loading={saveMutation.isPending}
-            onClick={saveDraft}
-            variant="primary"
-          >
+          <AsyncButton icon={<Save aria-hidden className="button__icon" />} loading={saveMutation.isPending} onClick={saveDraft} variant="primary">
             {t("common.save")}
           </AsyncButton>
           <Button
@@ -971,6 +655,7 @@ export function BackgroundManagerPage() {
         </aside>
 
         <section className="settings-grid">
+          {/* Info section */}
           <section className="section">
             <div className="section__header">
               <h2 className="section__title">{t("background.section.info")}</h2>
@@ -980,11 +665,7 @@ export function BackgroundManagerPage() {
                   loading={translateMutation.isPending}
                   onClick={() => {
                     if (!draft.name.trim() && !draft.bg_tags.trim() && !draft.bgm_tags.trim()) {
-                      showToast({
-                        kind: "error",
-                        message: t("common.fixInvalidFields"),
-                        title: t("common.validationFailed"),
-                      });
+                      showToast({ kind: "error", message: t("common.fixInvalidFields"), title: t("common.validationFailed") });
                       return;
                     }
                     translateMutation.mutate();
@@ -997,11 +678,7 @@ export function BackgroundManagerPage() {
                   icon={<Trash2 aria-hidden className="button__icon" />}
                   onClick={() => {
                     if (!currentBackgroundName) {
-                      showToast({
-                        kind: "error",
-                        message: t("background.error.deleteFallback"),
-                        title: t("common.deleteFailed"),
-                      });
+                      showToast({ kind: "error", message: t("background.error.deleteFallback"), title: t("common.deleteFailed") });
                       return;
                     }
                     setPendingDelete({ kind: "background", name: currentBackgroundName });
@@ -1027,15 +704,13 @@ export function BackgroundManagerPage() {
               <label className="field-row">
                 <span className="field-row__label">{t("background.field.spritePrefix")}</span>
                 <span className="field-row__control">
-                  <TextInput
-                    onChange={(event) => update("sprite_prefix", event.target.value)}
-                    value={draft.sprite_prefix}
-                  />
+                  <TextInput onChange={(event) => update("sprite_prefix", event.target.value)} value={draft.sprite_prefix} />
                 </span>
               </label>
             </div>
           </section>
 
+          {/* Tags section */}
           <section className="section">
             <div className="section__header">
               <h2 className="section__title">{t("background.section.tags")}</h2>
@@ -1050,11 +725,7 @@ export function BackgroundManagerPage() {
                     loading={imageTagsSaveMutation.isPending}
                     onClick={() => {
                       if (!currentBackgroundName) {
-                        showToast({
-                          kind: "error",
-                          message: t("background.validation.nameRequired"),
-                          title: t("background.action.saveImageTags"),
-                        });
+                        showToast({ kind: "error", message: t("background.validation.nameRequired"), title: t("background.action.saveImageTags") });
                         return;
                       }
                       imageTagsSaveMutation.mutate();
@@ -1074,11 +745,7 @@ export function BackgroundManagerPage() {
                     loading={bgmTagsSaveMutation.isPending}
                     onClick={() => {
                       if (!currentBackgroundName) {
-                        showToast({
-                          kind: "error",
-                          message: t("background.validation.nameRequired"),
-                          title: t("background.action.saveBgmTags"),
-                        });
+                        showToast({ kind: "error", message: t("background.validation.nameRequired"), title: t("background.action.saveBgmTags") });
                         return;
                       }
                       bgmTagsSaveMutation.mutate();
@@ -1092,255 +759,73 @@ export function BackgroundManagerPage() {
             </div>
           </section>
 
-          <section className="section">
-            <div className="section__header">
-              <h2 className="section__title">{t("background.section.images")}</h2>
-              <div className="page__actions">
-                <AsyncButton
-                  icon={<Upload aria-hidden className="button__icon" />}
-                  loading={imageUploadMutation.isPending}
-                  onClick={() => {
-                    if (!currentBackgroundName) {
-                      showToast({
-                        kind: "error",
-                        message: t("background.validation.nameRequired"),
-                        title: t("background.asset.uploadImages"),
-                      });
-                      return;
-                    }
-                    if (!pendingImagePaths.length) {
-                      showToast({ kind: "error", title: t("background.asset.selectImages") });
-                      return;
-                    }
-                    imageUploadMutation.mutate();
-                  }}
-                  variant="ghost"
-                >
-                  {t("background.asset.uploadImages")}
-                </AsyncButton>
-                <Button
-                  icon={<Trash2 aria-hidden className="button__icon" />}
-                  onClick={() => {
-                    if (!currentBackgroundName || !draft.sprites.length) {
-                      showToast({ kind: "error", title: t("background.asset.clearImages") });
-                      return;
-                    }
-                    setPendingDelete({
-                      count: draft.sprites.length,
-                      kind: "all-images",
-                      name: currentBackgroundName,
-                    });
-                  }}
-                  variant="ghost"
-                >
-                  {t("background.asset.clearImages")}
-                </Button>
-              </div>
-            </div>
-            <div className="asset-editor">
-              <label className="field-row field-row--stack">
-                <span className="field-row__label">{t("background.asset.selectImages")}</span>
-                <span className="field-row__control">
-                  <FilePicker
-                    acceptedExtensions={[".gif", ".jpeg", ".jpg", ".png", ".webp"]}
-                    multiple
-                    onPathsChange={(paths) => {
-                      if (paths.length) {
-                        setPendingImagePaths(paths);
-                      }
-                    }}
-                    pickLabel={t("common.chooseFile")}
-                    pickerTitle={t("background.asset.selectImages")}
-                    value={
-                      pendingImagePaths.length
-                        ? t("background.asset.selectedFiles", { count: pendingImagePaths.length })
-                        : ""
-                    }
-                  />
-                </span>
-              </label>
-              {!draft.sprites.length ? <EmptyState title={t("background.asset.emptyImages")} /> : null}
-              {selectedImage ? (
-                <div className="asset-gallery-layout asset-gallery-layout--background">
-                  <ImageAssetGallery
-                    items={backgroundImageItems}
-                    onSelect={setSelectedImageIndex}
-                    selectedIndex={selectedImageIndex}
-                  />
-                  <aside className="asset-inspector">
-                    <div className="asset-inspector__preview asset-inspector__preview--background">
-                      {selectedImage.path ? (
-                        <img alt="" decoding="async" src={fileUrl(selectedImage.path)} />
-                      ) : (
-                        <ImageIcon aria-hidden className="asset-inspector__fallback" />
-                      )}
-                    </div>
-                    <label className="field-row field-row--stack">
-                      <span className="field-row__label">{t("background.asset.path")}</span>
-                      <span className="field-row__control">
-                        <PathDisplay className="path-display--input" path={selectedImage.path} />
-                      </span>
-                    </label>
-                    <label className="field-row field-row--stack">
-                      <span className="field-row__label">{t("background.asset.tag")}</span>
-                      <span className="field-row__control">
-                        <TextInput
-                          onChange={(event) => updateImageRowTag(selectedImageIndex, event.target.value)}
-                          value={imageRowTags[selectedImageIndex] ?? ""}
-                        />
-                      </span>
-                    </label>
-                    <div className="asset-inspector__actions">
-                      <AsyncButton
-                        icon={<Save aria-hidden className="button__icon" />}
-                        loading={imageTagsSaveMutation.isPending}
-                        onClick={() => {
-                          if (!currentBackgroundName) {
-                            showToast({
-                              kind: "error",
-                              message: t("background.validation.nameRequired"),
-                              title: t("background.action.saveImageTags"),
-                            });
-                            return;
-                          }
-                          imageTagsSaveMutation.mutate();
-                        }}
-                        variant="ghost"
-                      >
-                        {t("background.action.saveImageTags")}
-                      </AsyncButton>
-                      <AsyncButton
-                        icon={<Trash2 aria-hidden className="button__icon" />}
-                        loading={imageDeleteMutation.isPending}
-                        onClick={() => handleImageDelete(selectedImageIndex)}
-                        variant="ghost"
-                      >
-                        {t("common.remove")}
-                      </AsyncButton>
-                    </div>
-                  </aside>
-                </div>
-              ) : null}
-            </div>
-          </section>
+          {/* Sprite gallery */}
+          <BackgroundSpriteGallery
+            currentBackgroundName={currentBackgroundName}
+            deletePending={imageDeleteMutation.isPending}
+            imageRowTags={imageRowTags}
+            onClearImages={() => {
+              if (!currentBackgroundName || !draft.sprites.length) {
+                showToast({ kind: "error", title: t("background.asset.clearImages") });
+                return;
+              }
+              setPendingDelete({ count: draft.sprites.length, kind: "all-images", name: currentBackgroundName });
+            }}
+            onDeleteImage={handleImageDelete}
+            onPendingImagePathsChange={setPendingImagePaths}
+            onSaveImageTags={() => {
+              if (!currentBackgroundName) {
+                showToast({ kind: "error", message: t("background.validation.nameRequired"), title: t("background.action.saveImageTags") });
+                return;
+              }
+              imageTagsSaveMutation.mutate();
+            }}
+            onSelectImage={setSelectedImageIndex}
+            onUpdateImageTag={updateImageRowTag}
+            onUploadImages={() => imageUploadMutation.mutate()}
+            pendingImagePaths={pendingImagePaths}
+            saveTagsPending={imageTagsSaveMutation.isPending}
+            selectedImageIndex={selectedImageIndex}
+            sprites={draft.sprites}
+            uploadPending={imageUploadMutation.isPending}
+          />
 
-          <section className="section">
-            <div className="section__header">
-              <h2 className="section__title">{t("background.section.bgm")}</h2>
-              <div className="page__actions">
-                <AsyncButton
-                  icon={<Upload aria-hidden className="button__icon" />}
-                  loading={bgmUploadMutation.isPending}
-                  onClick={() => {
-                    if (!currentBackgroundName) {
-                      showToast({
-                        kind: "error",
-                        message: t("background.validation.nameRequired"),
-                        title: t("background.asset.uploadBgm"),
-                      });
-                      return;
-                    }
-                    if (!pendingBgmPaths.length) {
-                      showToast({ kind: "error", title: t("background.asset.selectBgm") });
-                      return;
-                    }
-                    bgmUploadMutation.mutate();
-                  }}
-                  variant="ghost"
-                >
-                  {t("background.asset.uploadBgm")}
-                </AsyncButton>
-                <AsyncButton
-                  icon={<Trash2 aria-hidden className="button__icon" />}
-                  loading={bgmBatchDeleteMutation.isPending}
-                  onClick={() => {
-                    const validSelectedBgmIndexes = selectedBgmIndexes.filter(
-                      (index) => index >= 0 && index < draft.bgm_list.length,
-                    );
-                    if (!validSelectedBgmIndexes.length) {
-                      showToast({ kind: "error", title: t("background.asset.noSelectedBgm") });
-                      return;
-                    }
-                    setPendingDelete({
-                      count: validSelectedBgmIndexes.length,
-                      indexes: validSelectedBgmIndexes,
-                      kind: "selected-bgm",
-                      name: currentBackgroundName,
-                    });
-                  }}
-                  variant="ghost"
-                >
-                  {t("background.asset.deleteSelectedBgm")}
-                </AsyncButton>
-                <Button
-                  icon={<Trash2 aria-hidden className="button__icon" />}
-                  onClick={() => {
-                    if (!currentBackgroundName || !draft.bgm_list.length) {
-                      showToast({ kind: "error", title: t("background.asset.clearBgm") });
-                      return;
-                    }
-                    setPendingDelete({
-                      count: draft.bgm_list.length,
-                      kind: "all-bgm",
-                      name: currentBackgroundName,
-                    });
-                  }}
-                  variant="ghost"
-                >
-                  {t("background.asset.clearBgm")}
-                </Button>
-              </div>
-            </div>
-            <div className="asset-editor">
-              <label className="field-row field-row--stack">
-                <span className="field-row__label">{t("background.asset.selectBgm")}</span>
-                <span className="field-row__control">
-                  <FilePicker
-                    acceptedExtensions={[".flac", ".m4a", ".mp3", ".ogg", ".wav"]}
-                    multiple
-                    onPathsChange={(paths) => {
-                      if (paths.length) {
-                        setPendingBgmPaths(paths);
-                      }
-                    }}
-                    pickLabel={t("common.chooseFile")}
-                    pickerTitle={t("background.asset.selectBgm")}
-                    value={
-                      pendingBgmPaths.length
-                        ? t("background.asset.selectedFiles", { count: pendingBgmPaths.length })
-                        : ""
-                    }
-                  />
-                </span>
-              </label>
-              {!draft.bgm_list.length ? <EmptyState title={t("background.asset.emptyBgm")} /> : null}
-              {draft.bgm_list.length ? (
-                <BackgroundBgmRows
-                  allSelected={allBgmSelected}
-                  clearSelectionLabel={t("background.asset.clearBgmSelection")}
-                  deleting={bgmDeleteMutation.isPending}
-                  filenameLabel={t("background.asset.filename")}
-                  items={sortedBgmItems}
-                  indexLabel={t("background.asset.index")}
-                  onDelete={handleBgmDelete}
-                  onSort={toggleBgmSort}
-                  onTagChange={updateBgmRowTag}
-                  onToggleAllSelection={toggleAllBgmSelection}
-                  onToggleSelection={toggleBgmSelection}
-                  pathLabel={t("background.asset.path")}
-                  previewLabel={t("background.asset.preview")}
-                  removeLabel={t("common.remove")}
-                  rowTags={bgmRowTags}
-                  selectLabel={t("background.asset.select")}
-                  selectAllLabel={t("background.asset.selectAllBgm")}
-                  selectedIndexes={selectedBgmIndexSet}
-                  sortDirection={bgmSort.direction}
-                  sortKey={bgmSort.key}
-                  tagLabel={t("background.asset.tag")}
-                />
-              ) : null}
-            </div>
-          </section>
+          {/* BGM section */}
+          <BackgroundMusicSection
+            batchDeletePending={bgmBatchDeleteMutation.isPending}
+            bgmList={draft.bgm_list}
+            bgmRowTags={bgmRowTags}
+            currentBackgroundName={currentBackgroundName}
+            deletePending={bgmDeleteMutation.isPending}
+            onBatchDelete={() => {
+              const validIndexes = selectedBgmIndexes.filter((i) => i >= 0 && i < draft.bgm_list.length);
+              if (!validIndexes.length) {
+                showToast({ kind: "error", title: t("background.asset.noSelectedBgm") });
+                return;
+              }
+              setPendingDelete({ count: validIndexes.length, indexes: validIndexes, kind: "selected-bgm", name: currentBackgroundName });
+            }}
+            onClearAll={() => {
+              if (!currentBackgroundName || !draft.bgm_list.length) {
+                showToast({ kind: "error", title: t("background.asset.clearBgm") });
+                return;
+              }
+              setPendingDelete({ count: draft.bgm_list.length, kind: "all-bgm", name: currentBackgroundName });
+            }}
+            onDelete={handleBgmDelete}
+            onPendingBgmPathsChange={setPendingBgmPaths}
+            onSortToggle={toggleBgmSort}
+            onTagChange={updateBgmRowTag}
+            onToggleAllSelection={toggleAllBgmSelection}
+            onToggleSelection={toggleBgmSelection}
+            onUpload={() => bgmUploadMutation.mutate()}
+            pendingBgmPaths={pendingBgmPaths}
+            selectedBgmIndexSet={selectedBgmIndexSet}
+            sortDirection={bgmSort.direction}
+            sortKey={bgmSort.key}
+            sortedBgmItems={sortedBgmItems}
+            uploadPending={bgmUploadMutation.isPending}
+          />
         </section>
       </div>
